FIX: Make TestUtilities compile again by fixing graph construction

+++ src/libPoSy/tests/TestPoSyOptimiser.cpp
@@ -24,7 +24,7 @@ TEST_F(TestPoSyOptimiser, FailsAssertionOptimisingWhenUnready) {
 TEST_F(TestPoSyOptimiser, IsReadyOnceDataIsSet) {
     Properties p{};
     PoSyOptimiser optimiser{p};
-    animesh::Graph<std::shared_ptr<Surfel>,int> g;
+    animesh::Graph<std::shared_ptr<Surfel>,float> g;
     optimiser.set_data(g);
 
     optimiser.optimise_do_one_step();
